ci: don't force package update for R 3.6
The uncached setup-r step for R 3.6 takes much longer than the cached
(~25m vs ~7m).  This may be because RSPM is being set to an old URL
(to get a compatible pmxTools), and, as a result, the upgrade=TRUE
leads to many newer source packages being installed from CRAN.

Try with upgrade=FALSE so that pak will prefer binaries from RSPM as
long as a package version from the date-pinned RSPM URL meets any
constraints specified by mrgsolve.

(Another possibility is that the specified RSPM URL doesn't have
available binaries, in which case this change won't help.  If that's
the case, it's still fine to use upgrade=FALSE given the latest
package versions are tested in the other builds.)